feature: implement ImapStream on top of Tls+WS streams, add tests
In order to run the imap_stream (with wasm-pack), the ws-echo-server from
test-utils must be running as well.

The ws-echo-server intentionally uses OpenSSL for server emulation to ensure
interoperability between Rustls and OpenSSL (which will be 99.99% of TLS
implementations on the real IMAP servers that we will interact with.

Turns out that we don't need a special WsStream or TlsStream, because people
already did that for us, we just need to combine those and wrap them into
an ImapStream, that we will then be feeding into an ImapClient, once we
have it.